Update aro-operations-Management-and-BCDR.md
Included reference to OpenShift Logging
Including the link to cluster configuration requirements
MachineAutoScaler is already mentioned
Included a comment to review the disk types with zones. ZRS support is available with CSI driver. Need to check with PG is any caveats.
Did not make a mention of velero but called out backing up application and data pointing to the official docs based on velero
Mentioned canary rollout and added links
Good point regarding infra nodes that they may be charged in ARO. Should we retreat on the recommendation? Need to check with the PG on it?
CSI drivers work with ARO but need to check with PG on what the roadmap item means